pyproject.toml
[project]
name = "feedback-project"
version = "0.1.0"
description = ""
authors = [
    {name = "Harshjeet",email = "harshjeetkumar0@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "celery (>=5.5.2,<6.0.0)",
    "redis (>=6.0.0,<7.0.0)",
    "reportlab (>=4.4.0,<5.0.0)",
    "flower (>=2.0.1,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)"
]

[tool.poetry]
packages = [{include = "feedback_project", from = "src"}]


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
